Import dolfin_adjoint.* in dol.py if it has been imported before.
This approach is safe if dolfin_adjoint is not installed,
and will allow zero-configuration use of headflow with dolfin-adjoint
namespace in place of dolfin namespace everywhere.

However, the dolfin-adjoint specific functionality such as
disabling annotation will not work with only dolfin, so there is
still quite a few issues to figure out before this is seamless.