Improve implementation of penalty bcs.
Now the scheme is generalized so it can apply
Dirichlet BCs and penalty BCs for the pressure
from the same problem specification. To achieve
this, the NSScheme.fetch_bcs function is ignored
and an alternative version is implemented.
Also, the demands on the problem specific boundary
condition format are slightly different, so it's
not quite interchangeable (yet).