Fix pressure and velocity fields to handle mixed space.
This is a temporary hack, need a better solution.
Also added name= to the functions created here.
That will require a fairly new dolfin, but the mixed fixed
point code probably doesn't work with older dolfin anyway.

In general we should add name= to all Functions, to
get proper labels in paraview.