Add dummy script with installer.
Want something like:

  headflow-showcase <casedir>

showing an overview of the contents of casedir, including some parsing of metadata.
The metadata and casedir parsing utils needed for this will double up as utils
for testing that postprocessing works properly, and for writing scripts that
extend a casedir with additional postprocessing of the initially dumped data.