Make transient time dependencies work, validated by unit tests.
These latest changes also solves optimal copying of functions
for caching in postprocessing. Iff the plan is to cache a
Function, it is copied in get. NB! If a PPField.compute returns
something else than a Function, it will not be copied. Such
objects are assumed to be new objects each time.