Add fields Linfnorm and LocalCfl, see changes to demo_main.py for how to use.
Note 1: The LocalCfl field computes (sqrt(u**2) * dt / h) on each cell,
where h is the circumradius. This is not the actual definition of CFL,
so interpret with care!

Note 2: The Linfnorm of a Function is implemented as the Linfnorm of
the dof vector, which is not exactly the Linfnorm of functions in
spaces other than DG0.