Work on optimization demo.
Implemented most of the functionality needed for computing a
forward problem (an analytical womersley solution in 2D for now),
observing the output, and running a new forward problem with
the computed observations. Next step is controlling annotation
with dolfin-adjoint so we get to test replay of the second run.
Then we can test the gradient.